Introduce IPsec Policy Agent structure
This commmit introduces the structure for the IPsec Policy Agent
Kubernetes Operator. This operator running on each node is responsible
for monitoring the configmap that store IPsec configurations to protect
inter host pod to pod network traffic for that node. When the configmap
is created or updated by IPsec Policy Manager operator, the agent
operator detects the change and generates or updates the k8s-node.conf
file in the host's swanctl directory. The agent is responsible for the
sync of the configuration stored in the configmap and the file on the
host.

The base of the directories and files are generated by Operator SDK.
Reference: https://sdk.operatorframework.io/docs/

Test Plan:
PASS: Build and push images for IPsec Policy Agent Operator.
PASS: In a DX environment, install ipsec-policy-agent-operator using
      dist/install.yaml. Verify the following:
      - A new DaemonSet is introduced to the cluster for the policy
      agent along with their RBAC rules for this new operator.
      - The ipsec-policy-agent-operator image is present on system and
      for each node in the cluster a policy agent pod is created and on
      Running state.
      - Each node has its own /etc/swanctl/conf.d/k8s-node.conf file
      created by IPsec Policy Agent Operator.
      - When the configmap is updated, the k8s-node.conf file is
      updated accordingly.

Story: 2011127
Task: 51987

Co-Authored-By: Andy Ning <andy.ning@windriver.com>
Co-Authored-By: Leonardo Mendes <Leonardo.MendesSantana@windriver.com>
Change-Id: I5b8dbc2e2a9bf7106075a484e238dcf39e9c4d0b
Signed-off-by: Manoel Benedito Neto <Manoel.BeneditoNeto@windriver.com>