Add App Lifecycle for IPsec Policy Operator
This commit adds the python implementation that defines ipsec-policy-
operator app lifecycle. It is expected that the helm charts for IPsec
Policy Agent and Manager are installed along with this app, in order to
do that, the setup.cfg defines the entrypoints for the app helm app and
each helm chart plugin that will be installed.

The python implementation defines the App lifecycle and Helm App for
IPsec Policy Operator with a pre-apply rule to create the local
registry secret for the required images pulled for each chart.

This commit also adds the Zuul jobs for this python project.

Note: The execution of lifecycle hooks will be observed in an upcoming
review that configure and include fluxcd manifests for this app. The
link to this upcoming review is referenced below.
Reference: https://review.opendev.org/c/starlingx/app-ipsec-policy-operator/+/948485

Test Plan:
PASS: Build k8sapp_ipsec_policy_operator package.

Story: 2011127
Task: 51955

Depends-on: https://review.opendev.org/c/starlingx/app-ipsec-policy-operator/+/948481
Co-Authored-By: Andy Ning <andy.ning@windriver.com>
Co-Authored-By: Leonardo Mendes <Leonardo.MendesSantana@windriver.com>
Change-Id: I47ac1976e4784d680292d5eb011c20ebe8b545de
Signed-off-by: Manoel Benedito Neto <Manoel.BeneditoNeto@windriver.com>