Define app behavior and add app lables to nodes
The Daemonset that defines IPsec Policy Agent schedules a pod creation
for each cluster node that has `ipsec-policy-agent-operator: enabled`
label. As the IPsec Policy Agent pods are required to be running on
every node, this commit implements the addition of this label to every
node in the Kubernetes cluster through the application lifecycle.

During pre-apply and evaluate reapply operations the IPsec Policy Agent
is addressed to each node present or added to the system. If the app is
removed, the app labels and registry secrets once created are removed
and the app is uninstalled.

This commit defines the app behavior to trigger the evaluate reapply
operation and app desired state.

Test Plan:
PASS: Build ipsec-policy-operator packages.
PASS: Build a new image and deploy a DX system. Observe that
      ipsec-policy-operator is in uploaded state in the system.
PASS: Apply ipsec-policy-operator app and observe that every node in
      the Kubernetes Cluster has `ipsec-policy-agent-operator: enabled`
      label.
PASS: After ipsec-policy-operator app is installed, IPsec Policy Agent
      labels are present on Kubernetes cluster nodes, observe that an
      IPsec Policy Agent DaemonSet scheduled a pod to run in every
      node.
PASS: Add a new worker node in the system and observe that IPsec Policy
      Agent label is created for this node in sysinv database. After
      host is deployed, verify the presence of the label in Kubernetes
      node and that a new IPsec Policy Agent pod is scheduled to this.
PASS: Delete a node from system and observe that IPsec Policy Agent
      label created for this node is removed from label table.

Story: 2011127
Task: 52101

Depends-on: https://review.opendev.org/c/starlingx/app-ipsec-policy-operator/+/948485
Change-Id: I2edf33f8f92546d000b9eceee48fa68ef179202b
Signed-off-by: Manoel Benedito Neto <Manoel.BeneditoNeto@windriver.com>