Fix IPSec connection termination
This commit fixes the IPSec connection termination when
the last policy is deleted. The operator was just unloading
the connection in this situation, but it was observed that
in some cases, the SAs kept established even after unload
the connection. So, now the connections are terminated and
unloaded every time.

Test Plan:
PASS: Build an agent docker image and push to docker registry.
PASS: Build stx-ipsec-policy-operator package.
PASS: Install stx-ipsec-policy-operator using its tar.gz file via
      sysinv cli. The Operator structure is installed on Kubernetes
      cluster and pods are on Running State. Then apply a sample
      IPsec policy yaml file to create one policy CR. After that,
      delete the policy and observe that the configmaps and the
      k8s-nodes.conf file from all nodes are deleted and the
      connections are unloaded and the SAs are terminated correctly.

Story: 2011127
Task: 52369

Change-Id: If567408b97723e7a25ab8c121162265ee320b7d8
Signed-off-by: Leonardo Mendes <Leonardo.MendesSantana@windriver.com>