Confine operator pods to selected nodes
This udpate added nodeSelector to IPsec policy mananger deployment to
confine its pods to controller hosts only.

This update also added tolerations to manager deployment and agent
daemonset so that their pods can be scheduled on controller nodes. (in
standard system, there are taints on control plan nodes).

With this change, there will be only one manager pod running on one of
the controllers, and one agent pod running on each of the hosts.

Test Plan:
PASS: On a standard system (mutil nodes with workers), apply the
      ipsec-policy-operator app, verify that there is only one manager
      pod running on either one of the controllers, and there is one
      agent pod running each of the hosts.

Story: 2011127
Task: 52400

Change-Id: I0bcd148b88fbb3dc5412e85c817a070b06191185
Signed-off-by: Andy Ning <andy.ning@windriver.com>