Add a pre_remove hook in the IPSec app lifecycle
This commit adds a pre_remove hook in the IPSec app lifecycle to
delete all the configmaps generated by the manager operator in order
to allow the agent operator execute his cleanup routine in each node.
So, the connections are unloaded, SAs are terminated and no longer
present in the kernel, IPsec configuration file (k8s-node.conf) was
erased from each node, resulting in the complete removal of this app
from the system.

Note: pre_remove hook is under APP_LIFECYCLE_TYPE_SEMANTIC_CHECK
because at the moment is the only one that allow reject the operation
in case of error, the other ones, just log the error and continue the
procedure.

Test Plan:
PASS: On a DX system, install stx-ipsec-policy-operator using its
      tar.gz file via sysinv cli. The Operator structure is installed
      on Kubernetes cluster and pods are on Running State. Apply a
      sample IPsec policy yaml file and observe that the configmaps
      are created, the connections are loaded, but as the pod-to-pod
      IPsec child SA's start_action was changed from "start" to "trap",
      a traffic need to forced to establish the SA. After that, run
      "system application-remove ipsec-policy-operator" to remove the
      app and observe that the app was removed and the ipsec config
      file was deleted in each node, the connections were unloaded
      and the SAs were terminated.
PASS: On a DX system, force an error in configmap deletion, trigger
      application-remove command for ipsec-policy-operator app and
      observe the application remove failed with warning message:

      "Application-apply rejected: Application-remove rejected: Failed
      to delete configmap kube-root-ca.crt: forced error"

Story: 2011127
Task: 52470

Change-Id: I4b86643a9f6ef12012f9790ae09c8cecb79f9667
Signed-off-by: Leonardo Mendes <Leonardo.MendesSantana@windriver.com>